refactor: svg로부터 직접 import
