+++ pages/side-panel/src/components/LoginSection.tsx
@@ -1,12 +1,11 @@
 import { WEB_URL } from '@extension/shared/constants';
 import { Tab } from '@extension/shared/utils/extension';
-import { TopRightArrow } from '@src/icons';
+import TopRightArrow from '../../public/svgs/top_right_arrow.svg';
 
 export default function LoginSection() {
   return (
     <div className="flex h-full flex-col items-center justify-center">
       <p>메모 기능을 이용하려면 로그인이 필요합니다.</p>
-
       <button
         type="button"
         className="flex cursor-pointer items-center gap-2"
