+++ pages/side-panel/src/components/MemoHeader.tsx
@@ -1,7 +1,8 @@
-import { useDidMount, useTabQuery, useUserPreferDarkMode } from '@extension/shared/hooks';
-import { TopRightArrow } from '../icons';
 import { WEB_URL } from '@extension/shared/constants';
+import { useDidMount, useTabQuery, useUserPreferDarkMode } from '@extension/shared/hooks';
 import { I18n, responseUpdateSidePanel, Tab } from '@extension/shared/utils/extension';
+import Plus from '../../public/svgs/plus.svg';
+import TopRightArrow from '../../public/svgs/top_right_arrow.svg';
 
 export default function MemoHeader() {
   const { data: tab, refetch: refetchTab } = useTabQuery();
@@ -27,6 +28,7 @@ export default function MemoHeader() {
         onClick={handleMemoClick}
         className="cursor-pointer"
       />
+      <Plus width={20} height={20} className="cursor-pointer" fill={isUserPreferDarkMode ? 'black' : 'white'} />
       <span className="w-4" />
       <span className="label-text w-full truncate text-right">{tab?.title}</span>
     </div>
