--- pages/side-panel/src/icons/index.ts
@@ -1,4 +1,0 @@
-export { default as Refresh } from './Refresh';
-export { default as Option } from './Option';
-export { default as TopRightArrow } from './TopRightArrow';
-export { default as Plus } from './Plus';
